# Make Tuvok faster with jq checks