--- a/tests/test_module.py
+++ b/tests/test_module.py
@@ -12,24 +12,24 @@ class TestModule(object):
     def test_warns_if_module_isnt_pinned_git(self, caplog):
         file = 'tests/test_module/module_git_missingref.tf'
         with Wrap(self, [file], expect_exit=True):
-            assert ('Modules sourced from GitHub should be pinned FAIL in {}'.format(file)) in caplog.text
+            assert ('[FAIL] github_module_ref:Modules sourced from GitHub should be pinned:some_module'.format(file)) in caplog.text
 
     def test_passes_if_module_has_ref_git(self, caplog):
         file = 'tests/test_module/module_git_hasref.tf'
         with Wrap(self, [file], expect_exit=False):
-            assert ('Modules sourced from GitHub should be pinned PASS in {}'.format(file)) in caplog.text
+            assert ('[PASS] github_module_ref:Modules sourced from GitHub should be pinned:{}'.format(file)) in caplog.text
 
     def test_warns_if_module_isnt_pinned_github(self, caplog):
         file = 'tests/test_module/module_github_missingref.tf'
         with Wrap(self, [file], expect_exit=True):
-            assert ('Modules sourced from GitHub should be pinned FAIL in {}'.format(file)) in caplog.text
+            assert ('[FAIL] github_module_ref:Modules sourced from GitHub should be pinned:some_module:{}'.format(file)) in caplog.text
 
     def test_passes_if_module_has_ref_github(self, caplog):
         file = 'tests/test_module/module_github_hasref.tf'
         with Wrap(self, [file], expect_exit=False):
-            assert ('Modules sourced from GitHub should be pinned PASS in {}'.format(file)) in caplog.text
+            assert ('[PASS] github_module_ref:Modules sourced from GitHub should be pinned:{}'.format(file)) in caplog.text
 
     def test_passes_if_module_not_git(self, caplog):
         file = 'tests/test_module/module_notgit.tf'
         with Wrap(self, [file], expect_exit=False):
-            assert ('Modules sourced from GitHub should be pinned FAIL in {}'.format(file)) not in caplog.text
+            assert ('[FAIL] github_module_ref:Modules sourced from GitHub should be pinned:some_module:{}'.format(file)) not in caplog.text