--- a/tests/test_module.py
+++ b/tests/test_module.py
@@ -32,7 +32,7 @@ class TestModule(object):
     def test_passes_if_module_not_git(self, caplog):
         file = 'tests/test_module/module_notgit.tf'
         with Wrap(self, [file], expect_exit=False):
-            assert ('Modules sourced from GitHub should be pinned FAIL in {}'.format(file)) not in caplog.text
+            assert ('[PASS] github_module_ref:Modules sourced from GitHub should be pinned:{}'.format(file)) in caplog.text
 
     def test_fails_if_module_has_rackspace_http_ref(self, caplog):
         file = 'tests/test_module/module_github_has_rackspace_ref.tf'
@@ -42,7 +42,7 @@ class TestModule(object):
     def test_passes_if_module_has_rackspace_ssh_ref(self, caplog):
         file = 'tests/test_module/module_git_has_rackspace_ref.tf'
         with Wrap(self, [file], expect_exit=False):
-            assert ('github_rackspace_module_use_ssh:Rackspace module references should use SSH source paths:{}'.format(file)) in caplog.text
+            assert ('[PASS] github_rackspace_module_use_ssh:Rackspace module references should use SSH source paths:{}'.format(file)) in caplog.text
 
     def test_warns_if_module_has_http_ref(self, caplog):
         file = 'tests/test_module/module_github_hasref.tf'