--- a/tuvok/checks/__init__.py
+++ b/tuvok/checks/__init__.py
@@ -1,6 +1,6 @@
 from enum import Enum
 
-from .base import BaseTuvokCheck, Severity
+from .base import BaseTuvokCheck, Severity, CheckResult
 from .jq import JqCheck
 from .null import NullCheck