Experimenting with removing the shell here. It turns out it wasn't very helpful, as the number of processes for a large repo, Running a thread pool helped a ton, though I had to build a class to hold all the relevant results from a check, so it was easier to hand around.

Before:
```
➜  tuvok git:(master) time tuvok ~/src/test_repo -V 2>&1 | grep -v tuvok
0.00s user 0.01s system 0% cpu 23.164 total
```

After:
```
➜  tuvok git:(martinb3/faster_jq) time tuvok ~/src/test_repo -V 2>&1 | grep -v tuvok
0.00s user 0.01s system 0% cpu 8.434 total
```